Extract interface from CommandLineActionFactory
