+++ subprojects/launcher/src/test/groovy/org/gradle/launcher/MainTest.groovy
@@ -15,13 +15,14 @@
  */
 package org.gradle.launcher
 
-import spock.lang.Specification
-import org.junit.Rule
-import org.gradle.util.RedirectStdOutAndErr
 import org.gradle.api.Action
-import org.gradle.launcher.cli.CommandLineActionFactory
-import org.gradle.launcher.bootstrap.ExecutionListener
 import org.gradle.launcher.bootstrap.ExecutionCompleter
+import org.gradle.launcher.bootstrap.ExecutionListener
+import org.gradle.launcher.cli.CommandLineActionFactory
+import org.gradle.launcher.cli.DefaultCommandLineActionFactory
+import org.gradle.util.RedirectStdOutAndErr
+import org.junit.Rule
+import spock.lang.Specification
 
 class MainTest extends Specification {
     
@@ -36,7 +37,7 @@ class MainTest extends Specification {
     def actionFactoryImpl
     
     void actionFactory(Closure closure) {
-        actionFactoryImpl = new CommandLineActionFactory() { Action<ExecutionListener> convert(List args) { closure(args) } }
+        actionFactoryImpl = new DefaultCommandLineActionFactory() { Action<ExecutionListener> convert(List args) { closure(args) } }
     }
     boolean completedSuccessfully
     boolean completedWithFailure
