+++ subprojects/launcher/src/test/groovy/org/gradle/launcher/cli/DefaultCommandLineActionFactoryTest.groovy
@@ -38,7 +38,7 @@ import org.gradle.util.SetSystemProperties
 import org.junit.Rule
 import spock.lang.Specification
 
-class CommandLineActionFactoryTest extends Specification {
+class DefaultCommandLineActionFactoryTest extends Specification {
     @Rule
     public final RedirectStdOutAndErr outputs = new RedirectStdOutAndErr();
     @Rule
@@ -50,7 +50,7 @@ class CommandLineActionFactoryTest extends Specification {
     final LoggingManagerInternal loggingManager = Mock()
     final CommandLineAction actionFactory1 = Mock()
     final CommandLineAction actionFactory2 = Mock()
-    final CommandLineActionFactory factory = new CommandLineActionFactory() {
+    final CommandLineActionFactory factory = new DefaultCommandLineActionFactory() {
         @Override
         LoggingServiceRegistry createLoggingServices() {
             return loggingServices
